Make event::emit's generic private. Fix private transfer. Add tests
- Make event::emit's generic require a struct private
- Fix bug in the "private transfer" check
- Added tests for transfer and event
