+++ crates/sui-verifier-transactional-tests/tests/private_generics/no_public_transfer.exp
@@ -1,0 +1,41 @@
+processed 11 tasks
+
+task 1 'publish'. lines 8-11:
+created: object(103)
+written: object(102)
+
+task 2 'publish'. lines 13-18:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 3 'publish'. lines 20-29:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object_id'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 4 'publish'. lines 31-36:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::freeze_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 5 'publish'. lines 38-43:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::share_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 6 'publish'. lines 45-51:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 7 'publish'. lines 53-59:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 8 'publish'. lines 61-69:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 9 'publish'. lines 71-79:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_object'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 10 'publish'. lines 81-89:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_address'. Invalid transfer of object of type 'a::m::S'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
