+++ crates/sui-verifier-transactional-tests/tests/private_generics/no_public_transfer_generic.exp
@@ -1,0 +1,37 @@
+processed 10 tasks
+
+task 1 'publish'. lines 10-15:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 2 'publish'. lines 17-26:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object_id'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 3 'publish'. lines 28-33:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::freeze_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 4 'publish'. lines 35-40:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::share_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 5 'publish'. lines 42-48:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 6 'publish'. lines 50-56:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::transfer::transfer_to_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 7 'publish'. lines 58-65:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 8 'publish'. lines 67-74:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_object'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
+
+task 9 'publish'. lines 76-83:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::transfer_child_to_object. Invalid call to 'sui::transfer::transfer_child_to_address'. Invalid transfer of object of type 'T0'. The transferred object's type must be defined in the current module, or must have the 'store' type ability") } }
