+++ crates/sui-verifier-transactional-tests/tests/private_generics/private_event_emit.exp
@@ -1,0 +1,21 @@
+processed 6 tasks
+
+task 1 'publish'. lines 8-11:
+created: object(103)
+written: object(102)
+
+task 2 'publish'. lines 13-18:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::event::emit'. Invalid event type 'a::m::S'. The event's type must be defined in the current module") } }
+
+task 3 'publish'. lines 20-25:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::event::emit'. Invalid event type 'T0'. The event's type must be defined in the current module") } }
+
+task 4 'publish'. lines 27-32:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::event::emit'. Invalid event type 'u64'. The event's type must be defined in the current module") } }
+
+task 5 'publish'. lines 34-40:
+Error: Transaction Effects Status: MiscellaneousError
+Execution Error: ExecutionError: ExecutionError { inner: ExecutionErrorInner { kind: ModuleVerificationFailure, source: Some("_::m::t. Invalid call to 'sui::event::emit'. Invalid event type 'vector<_::m::X>'. The event's type must be defined in the current module") } }
