+++ crates/sui-verifier-transactional-tests/tests/private_generics/public_child_ref_functions.exp
@@ -1,0 +1,13 @@
+processed 4 tasks
+
+task 1 'publish'. lines 8-11:
+created: object(103)
+written: object(102)
+
+task 2 'publish'. lines 13-23:
+created: object(105)
+written: object(104)
+
+task 3 'publish'. lines 25-35:
+created: object(107)
+written: object(106)
