+++ crates/sui-verifier-transactional-tests/tests/private_generics/public_transfer_with_store.exp
@@ -1,0 +1,41 @@
+processed 11 tasks
+
+task 1 'publish'. lines 9-12:
+created: object(103)
+written: object(102)
+
+task 2 'publish'. lines 14-19:
+created: object(105)
+written: object(104)
+
+task 3 'publish'. lines 21-29:
+created: object(107)
+written: object(106)
+
+task 4 'publish'. lines 31-36:
+created: object(109)
+written: object(108)
+
+task 5 'publish'. lines 38-43:
+created: object(111)
+written: object(110)
+
+task 6 'publish'. lines 45-51:
+created: object(113)
+written: object(112)
+
+task 7 'publish'. lines 53-59:
+created: object(115)
+written: object(114)
+
+task 8 'publish'. lines 61-69:
+created: object(117)
+written: object(116)
+
+task 9 'publish'. lines 71-79:
+created: object(119)
+written: object(118)
+
+task 10 'publish'. lines 81-89:
+created: object(121)
+written: object(120)
