+++ crates/sui-verifier/src/lib.rs
@@ -7,7 +7,7 @@ pub mod entry_points_verifier;
 pub mod global_storage_access_verifier;
 pub mod id_immutable_verifier;
 pub mod id_leak_verifier;
-pub mod private_transfer;
+pub mod private_generics;
 pub mod struct_with_key_verifier;
 
 use move_binary_format::{
